Add mapping on initial sync on consumer
Due to a lack of understanding how things work, I forgot to add mapping
to the initial sync.
