+++ config/delta-consumer/eredienst-mandatarissen/delta-sync-dispatching.js
@@ -8,11 +8,10 @@ const {
   SLEEP_TIME_AFTER_FAILED_DB_OPERATION,
   INGEST_GRAPH,
 } = require('./config');
-const { batchedDbUpdate } = require('./utils');
+const { batchedDbUpdate, mapTriple } = require('./utils');
 const endpoint = BYPASS_MU_AUTH_FOR_EXPENSIVE_QUERIES
   ? DIRECT_DATABASE_ENDPOINT
   : process.env.MU_SPARQL_ENDPOINT; //Defaults to mu-auth
-const map = require('./mappings.js');
 
 /**
  * Dispatch the fetched triples to a target graph.
@@ -86,37 +85,7 @@ async function processDeletes(lib, data) {
   );
 }
 
-/*
- * Maps the predicate or object of a triple onto a different value. This can be
- * used to make corrections (or deliberate mistakes to adhere to a different
- * version of a model ;) ) to triples before inserting them into a temporary
- * graph.
- *
- * @function
- * @param {Object} triple - A triple as given by the consumer. @see dispatch
- * @returns {Object} Same triple structure, but the predicate or object might
- * have changed.
- */
-function mapTriple(triple) {
-  //See if the class definition needs to be transformed
-  if (
-    triple.predicate === '<http://www.w3.org/1999/02/22-rdf-syntax-ns#type>' ||
-    triple.predicate === 'a'
-  ) {
-    //Transform the class (object)
-    const replacement = map.mappingsClasses.find(
-      (elem) => elem.original === triple.object
-    );
-    if (replacement) triple.object = replacement.loket;
-  } else {
-    //Transform the predicate
-    const replacement = map.mappingsPredicates.find(
-      (elem) => elem.original === triple.predicate
-    );
-    if (replacement) triple.predicate = replacement.loket;
-  }
-  return triple;
-}
+
 
 module.exports = {
   dispatch,
