+++ config/delta-consumer/eredienst-mandatarissen/utils.js
@@ -1,3 +1,37 @@
+const map = require('./mappings.js');
+
+/*
+ * Maps the predicate or object of a triple onto a different value. This can be
+ * used to make corrections (or deliberate mistakes to adhere to a different
+ * version of a model ;) ) to triples before inserting them into a temporary
+ * graph.
+ *
+ * @function
+ * @param {Object} triple - A triple as given by the consumer. @see dispatch
+ * @returns {Object} Same triple structure, but the predicate or object might
+ * have changed.
+ */
+function mapTriple(triple) {
+  //See if the class definition needs to be transformed
+  if (
+    triple.predicate === '<http://www.w3.org/1999/02/22-rdf-syntax-ns#type>' ||
+    triple.predicate === 'a'
+  ) {
+    //Transform the class (object)
+    const replacement = map.mappingsClasses.find(
+      (elem) => elem.original === triple.object
+    );
+    if (replacement) triple.object = replacement.loket;
+  } else {
+    //Transform the predicate
+    const replacement = map.mappingsPredicates.find(
+      (elem) => elem.original === triple.predicate
+    );
+    if (replacement) triple.predicate = replacement.loket;
+  }
+  return triple;
+}
+
 async function batchedDbUpdate(
   muUpdate,
   graph,
@@ -68,4 +102,5 @@ async function dbOperationWithRetry(
 
 module.exports = {
   batchedDbUpdate,
+  mapTriple,
 };
